Move read_topspin reader to external plugin package
- Extract read_topspin.py from core readers into plugins/spectrochempy-topspin/
- Replace core read_topspin.py with stub raising MissingPluginError
- Fix PluginManager.discover() to instantiate entry points and call register()
- Add plugin_manager.discover() call before plugin reader checks in __getattr__
- Add plugin reader fallback in NDDataset.__getattr__ for dataset method access
- Remove read_topspin from core readers/__init__.py, api_methods.py, dataset_methods.py
