+++ src/spectrochempy/lazyimport/api_methods.py
@@ -177,7 +177,6 @@ _LAZY_IMPORTS = {
     "read_sdr": "spectrochempy.core.readers.read_soc",
     "read_hdr": "spectrochempy.core.readers.read_soc",
     "read_spc": "spectrochempy.core.readers.read_spc",
-    "read_topspin": "spectrochempy.core.readers.read_topspin",
     "read_wdf": "spectrochempy.core.readers.read_wire",
     "read_wire": "spectrochempy.core.readers.read_wire",
     "read_zip": "spectrochempy.core.readers.read_zip",
